unify all single node tests into one single class
